prevent dependabot deploys (#75)
dependabot cannot share action secrets, the action secrets are required for deploying to environments (dev and test). If dependabot opens a PR then this fix will now prevent deployments to dev. This approach relies on testing to catch any problems with a package bump - when merged the journey will be tested in our test environment before released to higher environments.