Prevent SYSTEM_ROLES override
Prevent SYSTEM_ROLES to be removed when login.

At the moment when we login we refresh the account role in the database
with the roles we get from AD.

This prevents that. We check the system roles the user had, and adds
them back to the array of roles before we trigger the update.