FSPT-156: fix fonts and favicon being served
Currently the authenticator app services web pages that fallback on the
`arial` font and don't have the GOV.UK crown as the Favicon. This gives
in inconsistent user experience, looks visually broken and doesn't
follow the guidance to use "GDS Transport" on gov.uk domains.

Tell the flask app to serve static assets from the `/assets` path. As
we're not customising our styles from the base design system SCSS we'll
use the default path included in the pre-built CSS.

The relative paths for the `/fonts` and `/images` will now line up and
the appropriate font should be used.

Remove custom logic for manipulating the asset path in the CSS thats
copied over, that is error prone as evidenced by the fact it would
update for the font files but not the favicon. (Instead we adjust where
static files are served from Flask).