Remove connexion
https://mhclgdigital.atlassian.net/browse/FSPT-189

In the process of merging authenticator with pre-award-frontend, we
found that the versions of Flask and Connexion in use between everything
are incompatible.

We remove connexion here because it takes us in the general direction of
travel (everything is just a Flask app), and allows us to integrate the
authenticator code (after this change) with pre-award-frontend (which
doesn't use connexion currently). This feels better than adding
connexion to pre-award-frontend, to then remove it again later.